% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement.funct.R
\name{densi.curv.cop}
\alias{densi.curv.cop}
\title{Compute the density of level curves for copulae models}
\usage{
densi.curv.cop(lines, copi)
}
\arguments{
\item{lines}{location of the objective level curve for which the deesnity needs to be estimated in the 2D space}

\item{copi}{a copula function with the parameters fitted to the bivariate dataset}
}
\value{
density for each points (couple x,y) along the level curves for copulae
}
\description{
Based on the density function of selected copula
}
