% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit.curves.R
\name{digit.curves.p}
\alias{digit.curves.p}
\title{Creates homogeneous level curve for every model (adapted from the function digit.curves
of the geomorph package)}
\usage{
digit.curves.p(start, curve, nPoints, closed = TRUE)
}
\arguments{
\item{start}{A numeric vector of x,y, coordinates for the landmark defining the start of the curve}

\item{curve}{A matrix (p x k) of 2D coordinates for a set of ordered points defining a curve}

\item{nPoints}{Numeric how many semilandmarks to place equidistantly along the curve (not counting beginning and end points)}

\item{closed}{Logical Whether the curve is closed (TRUE) or open (FALSE)}
}
\value{
Function returns a matrix of coordinates for nPoints equally spaced semilandmarks sampled along the curve
}
\description{
Creates homogeneous level curve for every model (adapted from the function digit.curves
of the geomorph package)
}
