% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digit.curves.R
\name{evenPts.p}
\alias{evenPts.p}
\title{basic function for spacing out curve points via linear interpolation (adapted from the function digit.curves
of the geomorph package). The main different is that curves are normalized to allow an intercomaprison of confidence scores
regardless of the input data.
used in digit.curves.p}
\usage{
evenPts.p(x, n)
}
\arguments{
\item{x, n}{numeric vectors}
}
\description{
basic function for spacing out curve points via linear interpolation (adapted from the function digit.curves
of the geomorph package). The main different is that curves are normalized to allow an intercomaprison of confidence scores
regardless of the input data.
used in digit.curves.p
}
