% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_kde.R
\name{JT.KDE.ap}
\alias{JT.KDE.ap}
\title{Fit bivariate joint tail model with Kernel density estimator (Adapted from Cooley et al., 2019)}
\usage{
JT.KDE.ap(
  u2,
  pbas,
  pobj,
  beta,
  vtau,
  devplot = F,
  kk,
  mar1,
  mar2,
  px,
  py,
  interh = NA
)
}
\arguments{
\item{u2}{Two column data frame}

\item{pbas}{joint return period to be modelled with a kde}

\item{pobj}{objective joint return period modelled with the joint tail model}

\item{beta}{smoothing parameter for the transition between asymptotic dependent and independent regimes near the axes}

\item{vtau}{estimate of the rank correlation between the two variables}

\item{devplot}{additional plots for development (significantly slows the function)}

\item{kk}{uniform margins of the original data}

\item{mar1}{Values of the first margin}

\item{mar2}{Values of the second margin}

\item{px}{Uniform values of the first margin with a mixed distribution (empirical below and gpd above a threshold)}

\item{py}{Uniform values of the second margin with a mixed distribution (empirical below and gpd above a threshold)}

\item{interh}{type of hazard interrelation '\code{comb}' for compound and '\code{casc}' for cascade,}
}
\value{
Estimates of the level curve with the selected return period, the base level curve and threshold dependent extremal dependence measures
}
\description{
Fit bivariate joint tail model with Kernel density estimator (Adapted from Cooley et al., 2019)
}
