% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement.funct.R
\name{Margins.mod}
\alias{Margins.mod}
\title{Compute uniform margins with Generalized Pareto Distribution above threshold}
\usage{
Margins.mod(tr1, tr2, u)
}
\arguments{
\item{tr1}{extreme threshold for first variable}

\item{tr2}{extreme threshold for second variable}

\item{u}{Two column data frame}
}
\value{
pseudo observations (uniform margins) with a mixed distribution (empirical below and gpd above a threshold)
}
\description{
Compute uniform margins with Generalized Pareto Distribution above threshold
}
