% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jt_kde.R
\name{Cond.mod.ap}
\alias{Cond.mod.ap}
\title{Fit bivariate conditional extremes model (from Heffernan and Tawn, 2004 and texmex R package)}
\usage{
Cond.mod.ap(u2, tr1, tr2, tsim, num.sim, pobj = 0.001, interh = "comb")
}
\arguments{
\item{u2}{Two column data frame}

\item{tr1}{extreme threshold for first variable}

\item{tr2}{extreme threshold for second variable}

\item{tsim}{Prediction quantile. The quantile of the conditioning variable above which it will be simulated for importance sampling based prediction (from texmex)}

\item{num.sim}{The number of simulated observations to be generated for prediction (from texmex)}

\item{pobj}{objective joint return period modelled with the conditional extremes model}

\item{interh}{type of hazard interrelation '\code{comb}' for compound and '\code{casc}' for cascade,}
}
\value{
Estimates of the level curve with the selected return period, simulated extreme data and threshold dependent extremal dependence measures
}
\description{
Fit bivariate conditional extremes model (from Heffernan and Tawn, 2004 and texmex R package)
}
