% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Joint.excurves.R
\name{JointExceedanceCurve}
\alias{JointExceedanceCurve}
\alias{JointExceedanceCurve.default}
\alias{JointExceedanceCurve.mexMC}
\alias{JointExceedanceCurve.predict.mex}
\title{Creates level curve for conditional extremes model (adapted from the function jointExceedanceCurve
of the texmex package)}
\usage{
JointExceedanceCurve(Sample, ExceedanceProb, ...)

\method{JointExceedanceCurve}{default}(Sample, ExceedanceProb, n = 50, x = NULL, ...)

\method{JointExceedanceCurve}{mexMC}(
  Sample,
  ExceedanceProb,
  n = 50,
  x = NULL,
  which = 1:2,
  ...
)

\method{JointExceedanceCurve}{predict.mex}(
  Sample,
  ExceedanceProb,
  n = 200,
  x = NULL,
  boots = F,
  which = 1:2,
  meth = "J",
  ...
)
}
\arguments{
\item{...}{Further aguments to be passed to methods}

\item{which}{Vector length two identifying which margins to use for joint exceedance curve estimation. Can be integer vector, giving column numbers of original data matrix, or character vector identifying variables by name (these must match column names in original data).}
}
\description{
Creates level curve for conditional extremes model (adapted from the function jointExceedanceCurve
of the texmex package)
}
